fix(agents): break final delivery type cycle
