+++ src/agents/subagent-registry.types.ts
@@ -1,6 +1,6 @@
 import type { DeliveryContext } from "../utils/delivery-context.types.js";
 import type { SubagentRunOutcome } from "./subagent-announce-output.js";
-import type { FinalDeliveryState } from "./subagent-final-delivery-state.js";
+import type { FinalDeliveryState } from "./subagent-final-delivery.types.js";
 import type { SubagentLifecycleEndedReason } from "./subagent-lifecycle-events.js";
 import type { SpawnSubagentMode } from "./subagent-spawn.types.js";
